requirements.txt
openai
python-dotenv
fastapi
uvicorn
python-multipart
pillow
pytesseract
